pyproject.toml
[tool.poetry]
name = "refresher"
version = "0.1.0"
description = "Refresher project to gather thoughts and try out packages, insallments and different setups"
authors = ["ibrahimroshdy <ibrahimabdelrahman@aucegypt.edu>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.2"
tqdm = "^4.64.0"
pyfiglet = "^0.8.post1"
fuzzywuzzy = "^0.18.0"
ftfy = "^6.1.1"
termcolor = "^1.1.0"
speedtest-cli = "^2.1.3"
django-import-export = "^2.8.0"
gunicorn = "^20.1.0"
loguru = "^0.6.0"
django-model-utils = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
